requirement.txt
tensorflow
tensorflow-hub
pillow
pymodbustcp
pyserial
wheel
opencv-contrib-python
numpy
matplotlib
scipy
pyzbar
pylibdmtx